+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - Add the ability to change the password for the "root" user during the first-time setup.
 - Add the ability to change the password for the LUKS storage encryption during the first-time setup on the secure image.
 - Add support for Framework laptops.
+- Change the fast Pacman mirror service to run as part of the first-time setup instead of the installation.
 
 ## 3.2.1 - The SteamOS 3.4 Update